+++ app/Http/Controllers/Api/EstudantesController.php
@@ -11,6 +11,8 @@ class EstudantesController extends Controller
 {
     public function index()
     {
+        $this->authorize('view', Estudante::class);
+
         return new EstudanteResource(
             Estudante::apiQuery()
         );
@@ -18,6 +20,8 @@ class EstudantesController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', Estudante::class);
+
         $dados = $request->validate([
             'nome' => 'required',
             'matricula' => 'required',
@@ -30,6 +34,8 @@ class EstudantesController extends Controller
 
     public function update(Estudante $estudante, Request $request)
     {
+        $this->authorize('edit', $estudante);
+
         $dados = $request->validate([
             'nome' => 'required',
             'matricula' => 'required',
@@ -42,9 +48,11 @@ class EstudantesController extends Controller
 
     public function destroy(Estudante $estudante)
     {
+        $this->authorize('delete', $estudante);
+
         try {
             $estudante->delete();
-            return response('Estudante removido.');
+            return response('Estudante removido.', 201);
         } catch (\Exception $exception) {
             return response(['message' => 'Estudante sendo referenciado em algum empréstimo.'], 422);
         }
